+++ collector-bitmovin-player/src/main/java/com/bitmovin/analytics/bitmovin/player/BitmovinPlayerCollector.java
@@ -5,8 +5,13 @@ import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.os.Build;
+import com.bitmovin.analytics.BitmovinAdAnalytics;
 import com.bitmovin.analytics.BitmovinAnalytics;
 import com.bitmovin.analytics.BitmovinAnalyticsConfig;
+import com.bitmovin.analytics.Collector;
+import com.bitmovin.analytics.DefaultCollector;
+import com.bitmovin.analytics.adapters.AdAdapter;
+import com.bitmovin.analytics.adapters.PlayerAdapter;
 import com.bitmovin.analytics.bitmovin.player.features.BitmovinFeatureFactory;
 import com.bitmovin.analytics.config.SourceMetadata;
 import com.bitmovin.analytics.data.DeviceInformationProvider;
@@ -14,8 +19,9 @@ import com.bitmovin.analytics.features.FeatureFactory;
 import com.bitmovin.player.api.Player;
 import com.bitmovin.player.api.source.Source;
 import java.util.HashMap;
+import org.jetbrains.annotations.NotNull;
 
-public class BitmovinPlayerCollector extends BitmovinAnalytics {
+public class BitmovinPlayerCollector extends DefaultCollector<Player> implements Collector<Player> {
     private HashMap<Source, SourceMetadata> sourceMetadataMap = new HashMap<>();
 
     /**
@@ -33,33 +39,35 @@ public class BitmovinPlayerCollector extends BitmovinAnalytics {
         this(bitmovinAnalyticsConfig, bitmovinAnalyticsConfig.getContext());
     }
 
-    public void attachPlayer(Player player) {
-        DeviceInformationProvider deviceInformationProvider =
-                new DeviceInformationProvider(context, getUserAgent(context));
-        FeatureFactory featureFactory =
-                new BitmovinFeatureFactory(bitmovinAnalyticsConfig, this, player, context);
-        BitmovinSdkAdapter adapter =
-                new BitmovinSdkAdapter(
-                        player,
-                        this.bitmovinAnalyticsConfig,
-                        deviceInformationProvider,
-                        this.playerStateMachine,
-                        featureFactory,
-                        sourceMetadataMap);
+    @NotNull
+    @Override
+    protected PlayerAdapter createAdapter(
+            Player player,
+            @NotNull BitmovinAnalytics analytics,
+            @NotNull DeviceInformationProvider deviceInformationProvider) {
+        FeatureFactory featureFactory = new BitmovinFeatureFactory(analytics, player);
+        return new BitmovinSdkAdapter(
+                player,
+                analytics.getConfig(),
+                deviceInformationProvider,
+                analytics.getPlayerStateMachine(),
+                featureFactory,
+                sourceMetadataMap);
+    }
 
-        this.attach(adapter);
-
-        if (this.adAnalytics != null) {
-            BitmovinSdkAdAdapter adAdapter = new BitmovinSdkAdAdapter(player, this.adAnalytics);
-            this.attachAd(adAdapter);
-        }
+    @NotNull
+    @Override
+    protected AdAdapter createAdAdapter(Player player, @NotNull BitmovinAdAnalytics adAnalytics) {
+        return new BitmovinSdkAdAdapter(player, adAnalytics);
     }
 
     public void addSourceMetadata(Source playerSource, SourceMetadata sourceMetadata) {
         sourceMetadataMap.put(playerSource, sourceMetadata);
     }
 
-    private String getUserAgent(Context context) {
+    @NotNull
+    @Override
+    protected String getUserAgent(Context context) {
         ApplicationInfo applicationInfo = context.getApplicationInfo();
         int stringId = applicationInfo.labelRes;
         String applicationName = "Unknown";