+++ collector-exoplayer-v1/src/main/java/com/bitmovin/analytics/exoplayer/features/ExoPlayerFeatureFactory.kt
@@ -1,8 +1,6 @@
 package com.bitmovin.analytics.exoplayer.features
 
-import android.content.Context
 import com.bitmovin.analytics.BitmovinAnalytics
-import com.bitmovin.analytics.BitmovinAnalyticsConfig
 import com.bitmovin.analytics.features.Feature
 import com.bitmovin.analytics.features.FeatureFactory
 import com.bitmovin.analytics.features.errordetails.ErrorDetailBackend
@@ -12,7 +10,7 @@ import com.bitmovin.analytics.license.FeatureConfigContainer
 import com.google.android.exoplayer2.ExoPlayer
 import com.google.android.exoplayer2.SimpleExoPlayer
 
-class ExoPlayerFeatureFactory(private val analyticsConfig: BitmovinAnalyticsConfig, private val analytics: BitmovinAnalytics, private val player: ExoPlayer, private val context: Context) : FeatureFactory {
+class ExoPlayerFeatureFactory(private val analytics: BitmovinAnalytics, private val player: ExoPlayer) : FeatureFactory {
     override fun createFeatures(): Collection<Feature<FeatureConfigContainer, *>> {
         val features = mutableListOf<Feature<FeatureConfigContainer, *>>()
         var httpRequestTracking: HttpRequestTracking? = null
@@ -20,8 +18,8 @@ class ExoPlayerFeatureFactory(private val analyticsConfig: BitmovinAnalyticsConf
             val requestTrackingAdapter = ExoPlayerHttpRequestTrackingAdapter(player, analytics.onAnalyticsReleasingObservable)
             httpRequestTracking = HttpRequestTracking(requestTrackingAdapter)
         }
-        val errorDetailsBackend = ErrorDetailBackend(analyticsConfig.config, context)
-        var errorDetailTracking = ErrorDetailTracking(context, analyticsConfig, analytics, errorDetailsBackend, httpRequestTracking, analytics.onErrorDetailObservable)
+        val errorDetailsBackend = ErrorDetailBackend(analytics.config.config, analytics.context)
+        var errorDetailTracking = ErrorDetailTracking(analytics.context, analytics.config, analytics, errorDetailsBackend, httpRequestTracking, analytics.onErrorDetailObservable)
         features.add(errorDetailTracking)
         return features
     }