+++ collector/src/main/java/com/bitmovin/analytics/DefaultCollector.kt
@@ -1,0 +1,69 @@
+package com.bitmovin.analytics
+
+import android.content.Context
+import com.bitmovin.analytics.adapters.AdAdapter
+import com.bitmovin.analytics.adapters.PlayerAdapter
+import com.bitmovin.analytics.data.CustomData
+import com.bitmovin.analytics.data.DeviceInformationProvider
+
+abstract class DefaultCollector<TPlayer>
+/**
+ * Bitmovin Analytics
+ *
+ * @param config {@link BitmovinAnalyticsConfig}
+ */
+protected constructor(config: BitmovinAnalyticsConfig, context: Context) {
+    private val analytics = BitmovinAnalytics(config, context)
+
+    /**
+     * Bitmovin Analytics
+     *
+     * @param config [BitmovinAnalyticsConfig]
+     */
+    @Deprecated("Please use {@link #BitmovinAnalytics(BitmovinAnalyticsConfig, Context)} and pass {@link Context} separately.")
+    protected constructor(config: BitmovinAnalyticsConfig) : this(config, config.context)
+
+    var customData: CustomData
+        get() = analytics.customData
+        set(value) { analytics.customData = value }
+
+    val impressionId: String
+        get() = analytics.impressionId
+
+    val config: BitmovinAnalyticsConfig
+        get() = analytics.config
+
+    protected abstract fun getUserAgent(context: Context): String
+    protected abstract fun createAdapter(player: TPlayer, analytics: BitmovinAnalytics, deviceInformationProvider: DeviceInformationProvider): PlayerAdapter
+
+    protected open fun createAdAdapter(player: TPlayer, adAnalytics: BitmovinAdAnalytics): AdAdapter? = null
+
+    fun attachPlayer(player: TPlayer) {
+        val context = analytics.context
+        val deviceInformationProvider = DeviceInformationProvider(context, getUserAgent(context))
+        val adapter = createAdapter(player, analytics, deviceInformationProvider)
+        analytics.attach(adapter)
+
+        if (analytics.adAnalytics != null) {
+            createAdAdapter(player, analytics.adAnalytics)?.let {
+                analytics.attachAd(it)
+            }
+        }
+    }
+
+    fun detachPlayer() {
+        analytics.detachPlayer()
+    }
+
+    fun setCustomDataOnce(customData: CustomData) {
+        analytics.setCustomDataOnce(customData)
+    }
+
+    fun addDebugListener(listener: BitmovinAnalytics.DebugListener) {
+        analytics.addDebugListener(listener)
+    }
+
+    fun removeDebugListener(listener: BitmovinAnalytics.DebugListener) {
+        analytics.removeDebugListener(listener)
+    }
+}