+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Error detail tracking feature
 - `castTech` field
 
+### Removed
+- Unintended public methods from Collector classes
+
 ## Development v1
 
 ### Fixed
@@ -26,6 +29,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Error detail tracking feature
 - `castTech` field
+
+### Removed
+- Unintended public methods from Collector classes
 
 ## v2.1.0